refactor: duplicated drawSegItem function
